Return ECClassId as Hex in Instance Queries (#221)
* return ECClassId as hex instead of the classname in instance queries

* move QueryJsonAdaptor to its own file

* remove unnecessary comments

* remove unnecessary include

---------
